requirements.txt
aiofiles
aiogram
httpx
numpy
openai
pydantic
pydantic-settings
python-dotenv
requests
soundfile

alembic
sqlalchemy[asyncio]
asyncpg

amplitude-analytics
amplitude-python